+++ cache/delete_timing.go
