+++ cache/redis/entity.go
