+++ cache/redis/entity_test.go
