+++ cache/redis/integrated_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/gcp-kit/datastore-cache-go/redis"
+	"github.com/gcp-kit/datastore-cache-go/cache/redis"
 	redigo "github.com/gomodule/redigo/redis"
 	"github.com/google/go-cmp/cmp"
 	"google.golang.org/genproto/googleapis/datastore/v1"
